+++ index.html
@@ -226,7 +226,9 @@
       </div>
     </section>
   </template>
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/util.js" type="module"></script>
+  <script src="js/data.js" type="module"></script>
+  <script src="js/main.js" type="module"></script>
+  <script src="js/functions.js" type="module"></script>
 </body>
 </html>
